backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
mysqlclient==2.2.8
python-dotenv==1.2.2
